+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-(Nothing is unreleased.)
+- Fix bug in custom layout when its A3 layer has character key
 
 ## [3.1](https://github.com/andy23512/keybr-cc-extension/releases/tag/3.1)